refactor and extract validations
