+++ src/components/Input/BudgetInput.tsx
@@ -1,7 +1,11 @@
-import { Input } from "@components/Input";
+import {
+  budgetInputFilter,
+  cleanDailyBudget,
+} from "@components/Input/validations";
 import { usePromotionContext } from "@context";
-import { currencyStringToInt } from "@utils/currency";
 import { h, FunctionalComponent } from "preact";
+
+import { Input } from ".";
 
 export const BudgetInput: FunctionalComponent<{
   dailyBudget: string;
@@ -16,40 +20,15 @@ export const BudgetInput: FunctionalComponent<{
 }) => {
   const { currency } = usePromotionContext();
 
-  const budgetInputFilter = (value: string) => {
-    const decimal = currency.decimalSeparator;
-    const exponent = currency.exponent;
-
-    if (decimal) {
-      const disallowedCharacters = new RegExp(`[^0-9\\${decimal}]`, "g");
-      // This regex is used to prevent more digits after the decimal than allowed
-      const afterDecimalRegex = new RegExp(
-        `(\\${decimal}[0-9]{0,${exponent}}).*`,
-        "g"
-      );
-      return value
-        .replace(disallowedCharacters, "")
-        .replace(afterDecimalRegex, "$1");
-    }
-
-    const disallowedCharacters = new RegExp(`[^0-9]`, "g");
-    return value.replace(disallowedCharacters, "");
-  };
-
   const onBudgetBlur = (event: FocusEvent) => {
     const target = event.target as HTMLInputElement;
-    const finalValue = cleanDailyBudget(target.value);
+    const finalValue = cleanDailyBudget(
+      target.value,
+      currency,
+      defaultDailyBudget,
+      formatCurrencyWithoutSymbol
+    );
     setDailyBudget(finalValue);
-  };
-
-  const cleanDailyBudget = (value: string) => {
-    let intValue = currencyStringToInt(value, currency);
-
-    if (intValue === 0) {
-      intValue = defaultDailyBudget;
-    }
-
-    return formatCurrencyWithoutSymbol(intValue / currency.divisor);
   };
 
   return (
@@ -58,7 +37,7 @@ export const BudgetInput: FunctionalComponent<{
         ? { before: currency.symbol }
         : { after: currency.symbol })}
       value={dailyBudget}
-      inputFilter={budgetInputFilter}
+      inputFilter={budgetInputFilter(currency)}
       onInput={setDailyBudget}
       onBlur={(event) => onBudgetBlur(event as unknown as FocusEvent)}
       required
